[Enhancement] Decoupling AlterTableClauseVisitor and AlterTableAnalyzer (#26214)
